requirements.txt
dotenv
requests
google-genai
openai
pytest
ruff
mypy
black
matplotlib
toml
pyyaml
fastapi
uvicorn
streamlit